HostFilesUpdate change to include local changes
Changed the process for updating the files downloaded from
pistar.uk. Allows the entire file to be overwritten by a file named
"<FILE>-override". Otherwise the file will be downloaded from pistar.uk
and will add "<FILE>-prepend" and "<FILE>-append" files to the
downloaded file if they exist. The final operation is to allow a hook
script to modify the file with non-trivial changes. The hook script if
found is named "<FILE>.sh". All of these files are located in
~pi-star/etc.

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>